Struct bowling_sim::frame::Frame

pub struct Frame { /* fields omitted */ }

Represents a single frame in a bowling game. Implements functions for updating and modifying a frame. Contains the frames current status and number of pins knocked down (score).

Implementations

impl Frame

pub fn new(status: FrameStatus) -> Result<Self, BowlingError>

Creates a new frame. status must be either FrameStatus::Open(2) or FrameStatus::Final(3). Returns an error for any other status.

pub fn final_frame() -> Self

Creates a new 'final' frame.

pub fn add_bonus(&mut self, pins: usize) -> Result<(), BowlingError>

Attempt to add bonus points to a frame. Returns an error if attempting to add bonus points more than once (spare) or twice (strike).

pub fn bowl(&mut self, pins: usize) -> Result<(), BowlingError>

Attempts to bowl on a frame. Returns an error if attempting to bowl on a frame more than twice (or thrice for the final frame). Also returns an error if attempting to knock down more pins than currently left in the frame.

pub fn pins_up(&self) -> Result<usize, BowlingError>

Identifies the number of pins available to be knocked down based on Frame state, as follows

If FrameState is Closed or Bonus, then no pins are available (this frame is over)

If FrameState is Open and there are two rolls remaining, then the frame is just starting and 10 pins are available

If FrameState is Open and there is one roll available, then the number of pins available is 10 - the pins knocked down on the first roll (current score)

If FrameState is Final and there are three rolls remaining, then the frame is just starting and 10 pins are available.

If FrameState is Final and there are two rolls remaining, then there are two possibilities. If the first roll was a strike (the score is 10), then all the pins are back up and 10 pins are available. If the first roll was not a strike (the score is not 10), then the number of pins available is 10 - the pins knocked down on the first roll.

If FrameState is Final and there is one roll remaining, then there are two possibilities. If the score is 10 (first two rolls made a spare) or 20 (two strikes), then all 10 pins are available. The other possibility is that the first roll was a strike and the second roll was not, in which case the number of pins available is 20 - 10 (for the strike) - the number of pins knocked down on the second roll.

pub fn status(&self) -> FrameStatus

Returns current status of this frame

pub fn score(&self) -> usize

Returns current score of this frame